Cargo: fix

+++ crates/vent-common/src/render/model/model_loader.rs
@@ -34,7 +34,7 @@ impl ModelLoader3D {
             indices.extend(mesh.faces.iter().flat_map(|face| face.0.iter().copied()));
 
             vertices.extend(mesh.vertices.iter().map(|vertex| Vertex3D {
-                _pos: [vertex.x as f32, vertex.y as f32, vertex.z as f32],
+                _pos: [vertex.x, vertex.y, vertex.z],
                 _tex_coord: [0.0, 0.0],
             }));
         }
